Change to Async Lambda Function
This adds a lightweight function that API Gateway calls synchronously.
This lightweight function in turn triggers an async invocation of the
webhook consumer.

This is to ensure that the CIS webhook notifier gets a prompt response
and doesn't wait on the webhook consumer to do all it's work before continuing.

This can't be done directly by API Gateway because it doesn't support
invoking Lambda Async in Proxy mode (which is what we use). As a consequence
we need a second Lambda function