Treat RequestEntityTooLargeException as client error
The reason we've been seeing an increase in lambda errors is because the
payload this lambda has been receiving is too large to forward.

So, we handle that and return a proper error code.

Future work will be required to reduce the payload we're receiving from the
webhook notifier.

Jira: IAM-1698